Update code for single end reads
The variable "remap_num" was added previously in the code for parsing paired-end data and tracks the number of reads to be remapped. This tracking is essential for writing reads to the "to.remap" fastq files in the correct format. This has now been updated in the code for single-end data.